Dont' use torch for generating splitting indices
